Fix `theft_random_bits_bulk` and autoshrinking of bulk buffers.
The size for theft_random_bits_bulk was getting truncated due to a
uint8_t for the `bit_count` argument in
`theft_autoshrink_bit_pool_random`.

Fix how dropping works for bulk requests - only advance the
`dst_offset` when copying, otherwise it ends up with lots
of zeroes.

In the `bulk_random_bits` test, check whether it was able to find the
absolute minimal case within 25 failures. (Increase the bloom filter
size a bit because of the bulk buffer autoshrinking state space.)

+++ src/theft_autoshrink.h
@@ -140,7 +140,7 @@ void theft_autoshrink_free_bit_pool(struct theft *t,
 void
 theft_autoshrink_bit_pool_random(struct theft *t,
     struct theft_autoshrink_bit_pool *pool,
-    uint8_t bit_count, bool save_request,
+    uint32_t bit_count, bool save_request,
     uint64_t *buf);
 
 void
